requirements.txt
googletrans
bs4
configargparse
torch
numpy
requests
tqdm
google-compute-engine
gensim
ruamel.yaml